fix some bad imports due to refactoring
