+++ lib/components/primitive-components/Trace/Trace.ts
@@ -33,14 +33,8 @@ import type { NetLabel } from "../NetLabel"
 import type { Port } from "../Port"
 import type { TraceHint } from "../TraceHint"
 import type { TraceI } from "./TraceI"
-import { getTraceLength } from "./compute-trace-length"
-import { createSchematicTraceCrossingSegments } from "./create-schematic-trace-crossing-segments"
-import { createSchematicTraceJunctions } from "./create-schematic-trace-junctions"
-import { getMaxLengthFromConnectedCapacitors } from "./get-max-length-from-connected-capacitors"
-import { getSchematicObstaclesForTrace } from "./get-obstacles-for-trace"
-import { getOtherSchematicTraces } from "./get-other-schematic-traces"
-import { getTraceDisplayName } from "./get-trace-display-name"
-import { pushEdgesOfSchematicTraceToPreventOverlap } from "./push-edges-of-schematic-trace-to-prevent-overlap"
+import { getMaxLengthFromConnectedCapacitors } from "./trace-utils/get-max-length-from-connected-capacitors"
+import { getTraceDisplayName } from "./trace-utils/get-trace-display-name"
 import { isRouteOutsideBoard } from "lib/utils/is-route-outside-board"
 import { getObstaclesFromCircuitJson } from "lib/utils/obstacles/getObstaclesFromCircuitJson"
 import { computeSchematicNetLabelCenter } from "lib/utils/schematic/computeSchematicNetLabelCenter"
