+++ lib/components/primitive-components/Trace/Trace_doInitialSchematicTraceRender.ts
@@ -12,11 +12,11 @@ import { getEnteringEdgeFromDirection } from "lib/utils/schematic/getEnteringEdg
 import { getStubEdges } from "lib/utils/schematic/getStubEdges"
 import type { NetLabel } from "../NetLabel"
 import type { Port } from "../Port"
-import { createSchematicTraceCrossingSegments } from "./create-schematic-trace-crossing-segments"
-import { createSchematicTraceJunctions } from "./create-schematic-trace-junctions"
-import { getSchematicObstaclesForTrace } from "./get-obstacles-for-trace"
-import { getOtherSchematicTraces } from "./get-other-schematic-traces"
-import { pushEdgesOfSchematicTraceToPreventOverlap } from "./push-edges-of-schematic-trace-to-prevent-overlap"
+import { createSchematicTraceCrossingSegments } from "./trace-utils/create-schematic-trace-crossing-segments"
+import { createSchematicTraceJunctions } from "./trace-utils/create-schematic-trace-junctions"
+import { getSchematicObstaclesForTrace } from "./trace-utils/get-obstacles-for-trace"
+import { getOtherSchematicTraces } from "./trace-utils/get-other-schematic-traces"
+import { pushEdgesOfSchematicTraceToPreventOverlap } from "./trace-utils/push-edges-of-schematic-trace-to-prevent-overlap"
 import { computeSchematicNetLabelCenter } from "lib/utils/schematic/computeSchematicNetLabelCenter"
 import { Trace } from "./Trace"
